refactor: simplify build and packaging of collapse behavior
