+++ packages/behaviour/collapse/bs-accordion.js
@@ -1,6 +1,6 @@
 
 import { LitElement, html } from 'lit-element';
-import { BsAccordionCss } from './bs-accordion.css.js';
+import { BsAccordionCss } from './css/bs-accordion.css.js';
 import { BsContentRebootCss } from '@lit-element-bootstrap/content';
 
 export class BsAccordion extends LitElement {
