+++ packages/behaviour/collapse/bs-collapsable.js
@@ -1,6 +1,6 @@
 
 import { LitElement, html } from 'lit-element';
-import { BsCollapsableCss } from './bs-collapsable.css.js';
+import { BsCollapsableCss } from './css/bs-collapsable.css.js';
 import { BsContentRebootCss } from '@lit-element-bootstrap/content';
 
 export class BsCollapsable extends LitElement {
