+++ packages/behaviour/collapse/bs-collapse-mixin.js
