+++ packages/behaviour/collapse/bs-collapse-multi-toggle.js
