+++ packages/behaviour/collapse/bs-collapse-toggle.js
