+++ packages/behaviour/collapse/css/bs-accordion.css.js
