+++ packages/behaviour/collapse/css/bs-collapsable.css.js
