+++ packages/behaviour/collapse/css/index.js
@@ -1,0 +1,3 @@
+
+export { BsAccordionCss } from './bs-accordion.css.js';
+export { BsCollapsableCss } from './bs-collapsable.css.js';
