+++ packages/behaviour/collapse/index.js
@@ -1,8 +1,6 @@
-
+export * from './css'
 export { BsAccordion } from './bs-accordion.js';
-export { BsAccordionCss } from './bs-accordion.css.js';
 export { BsCollapsable } from './bs-collapsable.js';
-export { BsCollapsableCss } from './bs-collapsable.css.js';
 export { BsCollapseMixin } from './bs-collapse-mixin.js';
 export { BsCollapseToggle } from './bs-collapse-toggle.js';
 export { BsCollapseMultiToggle } from './bs-collapse-multi-toggle.js';
