+++ packages/behaviour/collapse/package.json
@@ -2,8 +2,8 @@
   "name": "@lit-element-bootstrap/collapse",
   "version": "1.0.0-alpha.2",
   "description": "Bootstrap v4.1.3 collapse behaviour elements in LitElement",
-  "main": "src/index.js",
-  "module": "src/index.js",
+  "main": "index.js",
+  "module": "index.js",
   "type": "module",
   "license": "MIT",
   "author": "Nikolas Christou",
@@ -15,10 +15,23 @@
   "publishConfig": {
     "access": "public"
   },
+  "files": [
+    "index.js",
+    "index.d.ts",
+    "global.d.ts",
+    "bs-*.js",
+    "bs-*.d.ts",
+    "css/-*.js",
+    "css/*.d.ts",
+    "custom-elements.json"
+  ],
   "scripts": {
-    "pika:build": "pika build",
-    "analyze": "lit-analyzer src",
-    "test": "karma start --coverage"
+    "prepack": "yarn build",
+    "build": "npm run wca-analyze:json",
+    "lit-analyze": "lit-analyzer bs-*.js css/*.js --failFast",
+    "wca-analyze:json": "wca analyze . --outFile custom-elements.json",
+    "lint:prettier": "prettier \"./**/*.js\" --tab-width 4 --list-different",
+    "format:prettier": "prettier \"./**/*.js\" --tab-width 4 --write"
   },
   "dependencies": {
     "@lit-element-bootstrap/content": "^1.0.0-alpha.3",
@@ -26,32 +39,15 @@
     "lit-html": "^1.0.0"
   },
   "devDependencies": {
-    "@pika/pack": "^0.5.0",
-    "@pika/plugin-build-web": "^0.6.1",
-    "@pika/plugin-bundle-web": "^0.6.1",
-    "@pika/plugin-standard-pkg": "^0.6.1",
     "@open-wc/testing": "^2.3.4",
     "@open-wc/karma-esm": "^2.7.1",
     "@open-wc/testing-karma": "^3.1.38",
     "deepmerge": "^4.1.1",
     "karma": "^4.3.0",
     "sinon": "^7.5.0",
-    "lit-analyzer": "^1.1.8"
-  },
-  "@pika/pack": {
-    "pipeline": [
-      [
-        "@pika/plugin-standard-pkg"
-      ],
-      [
-        "@pika/plugin-build-web"
-      ],
-      [
-        "@pika/plugin-bundle-web",
-        {
-            "entrypoint": "unpkg"
-        }
-      ]
-    ]
+    "lit-analyzer": "^1.1.8",
+    "@open-wc/prettier-config": "^0.1.12",
+    "eslint": "^6.6.0",
+    "web-component-analyzer": "^0.1.21"
   }
 }
